Refactor imagevector package
