--- charts/images.go
@@ -1,11 +1,0 @@
-// SPDX-FileCopyrightText: 2021 SAP SE or an SAP affiliate company and Gardener contributors
-//
-// SPDX-License-Identifier: Apache-2.0
-
-package charts
-
-import _ "embed"
-
-// Images YAML contains the contents of the images.yaml file.
-//go:embed images.yaml
-var ImagesYAML string
